Type casts

+++ trunk/src/gmtmercmap.c
@@ -164,15 +164,17 @@ int main (int argc, char **argv)
 	/* Initializing new GMT session */
 	if ((API = GMT_Create_Session ("TEST", GMTAPI_GMTPSL)) == NULL) exit (EXIT_FAILURE);
 	options = GMT_Prep_Options (API, argc-1, argv+1);	if (API->error) return (API->error);	/* Set or get option list */
-	if (!options || options->option == GMTAPI_OPT_USAGE) exit (GMT_gmtmercmap_usage (API, GMTAPI_USAGE));	/* Return the usage message */
-	if (options->option == GMTAPI_OPT_SYNOPSIS) exit (GMT_gmtmercmap_usage (API, GMTAPI_SYNOPSIS));	/* Return the synopsis */
+	if (!options || options->option == GMTAPI_OPT_USAGE) 
+		exit ((int)GMT_gmtmercmap_usage (API, GMTAPI_USAGE));	/* Return the usage message */
+	if (options->option == GMTAPI_OPT_SYNOPSIS) 
+		exit ((int)GMT_gmtmercmap_usage (API, GMTAPI_SYNOPSIS));	/* Return the synopsis */
 
 	/* Parse the command-line arguments */
 
 	GMT = GMT_begin_module (API, "GMT_gmtmercmap", &GMT_cpy);		/* Save current state */
-	if (GMT_Parse_Common (API, "-VR", "KOPUXxYycnpt>", options)) Return (API->error);
+	if (GMT_Parse_Common (API, "-VR", "KOPUXxYycnpt>", options)) Return ((int)API->error);
 	Ctrl = New_gmtmercmap_Ctrl (GMT);	/* Allocate and initialize a new control structure */
-	if ((error = GMT_gmtmercmap_parse (API, Ctrl, options))) Return (error);
+	if ((error = GMT_gmtmercmap_parse (API, Ctrl, options))) Return ((int)error);
 
 	/*---------------------------- This is the gmtmercmap main code ----------------------------*/
